Kraken_trend_bot/requirements.txt
ccxt
pandas
numpy
pyyaml
python-dotenv
scikit-learn
joblib
torch
feedparser
optuna
matplotlib